Improved names for composables. Tidyups
