+++ packages/frontend-2/middleware/requireSsoEnabled.ts
@@ -1,5 +1,5 @@
 import { workspaceRoute } from '~/lib/common/helpers/route'
-import { useWorkspaceSsoPublic } from '~/lib/workspaces/composables/sso'
+import { useWorkspacePublicSsoCheck } from '~/lib/workspaces/composables/sso'
 
 /**
  * Used to validate that the workspace has SSO enabled, redirects to workspace page if not
@@ -8,7 +8,7 @@ export default defineNuxtRouteMiddleware(async (to) => {
   const workspaceSlug = to.params.slug as string
   if (!workspaceSlug) return
 
-  const { workspace } = useWorkspaceSsoPublic(workspaceSlug)
+  const { workspace } = useWorkspacePublicSsoCheck(workspaceSlug)
   if (!workspace.value?.ssoProviderName) {
     return navigateTo(workspaceRoute(workspaceSlug))
   }
